Introduce the Flower Enhancement Docs process (#1122)
To document motivations and decisions for larger changes or features to Flower, an Enhancement Proposal process is proposed. This process is heavily influenced by similar processes in the Python language (https://peps.python.org) as well as in Kubernetes (https://git.k8s.io/enhancements/keps).
The goal of this process is to provide insight and transparency on upcoming features that might change how users or developers have to interact with Flower. By storing proposals as part of Flower's repository, the community has an easy and accessible way to understand and influence these important changes to Flower.

Co-authored-by: Orlando Hohmeier <hello@orlandohohmeier.com>
Co-authored-by: Daniel J. Beutel <daniel@adap.com>